+++ src/main/java/dev/hilligans/ourcraft/World/NewWorldSystem/GlobalPaletteImpl.java
@@ -28,12 +28,12 @@ public class GlobalPaletteImpl implements ISubChunk {
     }
 
     @Override
-    public IBlockState getBlockState(long x, long y, long z) {
+    public IBlockState getBlockState(int x, int y, int z) {
         if(blockStates == null) {
             return Blocks.AIR.getDefaultState1();
         }
         try {
-            short b = blockStates[(int) ((x * width + y) * height + z)];
+            short b = blockStates[((x * width + y) * height + z)];
             return b == 0 ? Blocks.AIR.getDefaultState1() : Ourcraft.GAME_INSTANCE.BLOCK_STATES.get(b);
         } catch (Exception e) {
             // e.printStackTrace();
@@ -42,7 +42,7 @@ public class GlobalPaletteImpl implements ISubChunk {
     }
 
     @Override
-    public IBlockState setBlockState(long x, long y, long z, IBlockState blockState) {
+    public IBlockState setBlockState(int x, int y, int z, IBlockState blockState) {
         if(blockStates == null) {
             if(blockState.getBlock() != Blocks.AIR) {
                 blockStates = new short[width * height * width];
